Tabulation of MMRM Results

These functions can be used to produce tables from a fitted MMRM produced with fit_mmrm().

Usage

# S3 method for mmrm
as.rtable(x, type = c("fixed", "cov", "diagnostic"), ...)

h_mmrm_fixed(x, format = "xx.xxxx")

h_mmrm_cov(x, format = "xx.xxxx")

h_mmrm_diagnostic(x, format = "xx.xxxx")

# S3 method for mmrm
tidy(x, ...)

s_mmrm_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

a_mmrm_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))

s_mmrm_lsmeans_single(df)

a_mmrm_lsmeans_single(df)

summarize_lsmeans(
  lyt,
  arms = TRUE,
  ...,
  table_names = "lsmeans_summary",
  .stats = NULL,
  .formats = NULL,
  .indent_mods = NULL,
  .labels = NULL
)

Arguments

x

(mmrm)
the original MMRM fit object.

type

(string)
type of table which should be returned.

...

additional argument format for controlling the numeric format.

format

(string)
format for the numbers in the table.

df

(data frame)
data set containing all analysis variables.

.in_ref_col

(logical)
TRUE when working with the reference level, FALSE otherwise.

show_relative

should the "reduction" (control - treatment, default) or the "increase" (treatment - control) be shown for the relative change from baseline?

lyt

(layout)
input layout where analyses will be added to.

arms

(flag)
should treatment variable be considered when using summarize_lsmeans layout generating function.

table_names

(character)
this can be customized in case that the same vars are analyzed multiple times, to avoid warnings from rtables.

.stats

(character)
statistics to select for the table.

.formats

(named character or list)
formats for the statistics.

.indent_mods

(named integer)
indent modifiers for the labels.

.labels

(named character)
labels for the statistics (without indent).

Value

as.rtable.mmrm() returns the fixed effects, covariance estimate or diagnostic statistics tables.

Functions

as.rtable.mmrm: Produce simple MMRM tables via the generic as.rtable().

h_mmrm_fixed: Helper function to produce fixed effects table.

h_mmrm_cov: Helper function to produce a covariance matrix table.

h_mmrm_diagnostic: Helper function to produce a diagnostic statistics table.

tidy.mmrm: Helper method (for broom::tidy()) to prepare a data.frame from an mmrm object containing the least-squares means and contrasts.

s_mmrm_lsmeans: Statistics function which is extracting estimates from a tidied least-squares means data frame.

a_mmrm_lsmeans: Formatted Analysis function which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

s_mmrm_lsmeans_single: Statistics function which is extracting estimates from a tidied least-squares means data frame when ARM is not considered in the model.

a_mmrm_lsmeans_single: Formatted Analysis function (when ARM is not considered in the model) which can be further customized by calling rtables::make_afun() on it. It is used as afun in rtables::analyze().

summarize_lsmeans: Analyze function for tabulating least-squares means estimates from tidied mmrm results.

Examples

library(dplyr)
library(lme4)
#> Loading required package: Matrix

dat <- sleepstudy %>%
  mutate(
    group = factor(rep(c("A", "B"), length = nrow(sleepstudy))),
    days_grouped = cut(
      Days,
      breaks = stats::quantile(Days, probs = seq(0, 1, length = 5)),
      include.lowest = TRUE
    ),
    Subject = case_when(
      group == "A" ~ as.character(Subject),
      TRUE ~ as.character(as.numeric(as.character(Subject)) + 50)
    )
  ) %>%
  distinct_at(.vars = c("Subject", "days_grouped", "group"), .keep_all = TRUE)

result <- fit_mmrm(
  vars = list(
    response = "Reaction",
    covariates = c(),
    id = "Subject",
    arm = "group",
    visit = "days_grouped"
  ),
  data = dat,
  cor_struct = "compound-symmetry",
  parallel = TRUE
)
as.rtable(result, type = "cov", format = "xx.x")
#>           [0,2]    (2,4.5]   (4.5,7]   (7,9] 
#> —————————————————————————————————————————————
#> [0,2]     2531.4   1443.7    1443.7    1443.7
#> (2,4.5]   1443.7   2531.4    1443.7    1443.7
#> (4.5,7]   1443.7   1443.7    2531.4    1443.7
#> (7,9]     1443.7   1443.7    1443.7    2531.4

result_no_arm <- fit_mmrm(
  vars = list(
    response = "Reaction",
    covariates = c(),
    id = "Subject",
    visit = "days_grouped"
  ),
  data = dat,
  cor_struct = "compound-symmetry",
  parallel = TRUE
)
as.rtable(result_no_arm, type = "cov", format = "xx.x")
#>           [0,2]    (2,4.5]   (4.5,7]   (7,9] 
#> —————————————————————————————————————————————
#> [0,2]     2479.0   1399.3    1399.3    1399.3
#> (2,4.5]   1399.3   2479.0    1399.3    1399.3
#> (4.5,7]   1399.3   1399.3    2479.0    1399.3
#> (7,9]     1399.3   1399.3    1399.3    2479.0
df <- broom::tidy(result)
df_no_arm <- broom::tidy(result_no_arm)
s_mmrm_lsmeans(df[8, ], .in_ref_col = FALSE)
#> $n
#> [1] 18
#> 
#> $adj_mean_se
#> [1] 350.85122  11.85897
#> 
#> $adj_mean_ci
#> [1] 327.1922 374.5102
#> attr(,"label")
#> [1] "95% CI"
#> 
#> $diff_mean_se
#> [1] 14.22172 16.77111
#> 
#> $diff_mean_ci
#> [1] -19.23716  47.68060
#> attr(,"label")
#> [1] "95% CI"
#> 
#> $change
#> [1] -0.04224739
#> attr(,"label")
#> [1] "Relative Reduction (%)"
#> 
#> $p_value
#> [1] 0.3993832
#> 
s_mmrm_lsmeans_single(df_no_arm[4, ])
#> $n
#> [1] 36
#> 
#> $adj_mean_se
#> [1] 343.740364   8.298227
#> 
#> $adj_mean_ci
#> [1] 327.1965 360.2842
#> attr(,"label")
#> [1] "95% CI"
#> 

dat_adsl <- dat %>%
  select(Subject, group) %>%
  unique()
basic_table() %>%
  split_cols_by("group", ref_group = result$ref_level) %>%
  add_colcounts() %>%
  split_rows_by("days_grouped") %>%
  summarize_lsmeans(show_relative = "increase") %>%
  build_table(
    df = broom::tidy(result),
    alt_counts_df = dat_adsl
  )
#>                                               A                    B         
#>                                             (N=18)               (N=18)      
#> —————————————————————————————————————————————————————————————————————————————
#> [0,2]                                                                        
#>   n                                           18                   18        
#>   Adjusted Mean (SE)                   256.652 (11.859)     264.496 (11.859) 
#>     95% CI                            (232.993, 280.311)   (240.837, 288.155)
#>   Difference in Adjusted Means (SE)                          7.844 (16.771)  
#>     95% CI                                                 (-25.615, 41.303) 
#>     Relative Increase (%)                                         3.1%       
#>     p-value (MMRM)                                               0.6415      
#> (2,4.5]                                                                      
#>   n                                           18                   18        
#>   Adjusted Mean (SE)                   288.649 (11.859)     282.992 (11.859) 
#>     95% CI                            (264.990, 312.308)   (259.333, 306.651)
#>   Difference in Adjusted Means (SE)                         -5.657 (16.771)  
#>     95% CI                                                 (-39.116, 27.801) 
#>     Relative Increase (%)                                        -2.0%       
#>     p-value (MMRM)                                               0.7369      
#> (4.5,7]                                                                      
#>   n                                           18                   18        
#>   Adjusted Mean (SE)                   312.178 (11.859)     308.518 (11.859) 
#>     95% CI                            (288.519, 335.837)   (284.859, 332.177)
#>   Difference in Adjusted Means (SE)                         -3.660 (16.771)  
#>     95% CI                                                 (-37.119, 29.799) 
#>     Relative Increase (%)                                        -1.2%       
#>     p-value (MMRM)                                               0.8279      
#> (7,9]                                                                        
#>   n                                           18                   18        
#>   Adjusted Mean (SE)                   336.630 (11.859)     350.851 (11.859) 
#>     95% CI                            (312.971, 360.289)   (327.192, 374.510)
#>   Difference in Adjusted Means (SE)                         14.222 (16.771)  
#>     95% CI                                                 (-19.237, 47.681) 
#>     Relative Increase (%)                                         4.2%       
#>     p-value (MMRM)                                               0.3994      

basic_table() %>%
  split_rows_by("days_grouped") %>%
  summarize_lsmeans(arms = FALSE) %>%
  build_table(
    df = broom::tidy(result_no_arm),
    alt_counts_df = dat_adsl
  )
#>                             all obs      
#> —————————————————————————————————————————
#> [0,2]                                    
#>   n                            36        
#>   Adjusted Mean (SE)    260.574 (8.298)  
#>     95% CI             (244.030, 277.118)
#> (2,4.5]                                  
#>   n                            36        
#>   Adjusted Mean (SE)    285.821 (8.298)  
#>     95% CI             (269.277, 302.365)
#> (4.5,7]                                  
#>   n                            36        
#>   Adjusted Mean (SE)    310.348 (8.298)  
#>     95% CI             (293.804, 326.892)
#> (7,9]                                    
#>   n                            36        
#>   Adjusted Mean (SE)    343.740 (8.298)  
#>     95% CI             (327.196, 360.284)